PR Comment 수정
- 1회성 메소드 삭제
- S3 패키징 수정
