+++ src/main/java/com/justcodeit/moyeo/study/interfaces/resource/SkillController.java
@@ -1,8 +1,9 @@
 package com.justcodeit.moyeo.study.interfaces.resource;
 
-import com.justcodeit.moyeo.study.interfaces.dto.BaseResponse;
-import com.justcodeit.moyeo.study.interfaces.dto.s3.SkillService;
+import com.justcodeit.moyeo.study.application.skill.SkillService;
+import com.justcodeit.moyeo.study.interfaces.dto.SuccessRes;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -14,7 +15,7 @@ public class SkillController {
     private final SkillService skillService;
 
     @PostMapping
-    public BaseResponse uploadZip() {
-        return skillService.skillSave();
+    public ResponseEntity<SuccessRes<Boolean>> uploadZip() {
+        return ResponseEntity.ok(new SuccessRes<>(skillService.skillSaveAllFromS3()));
     }
 }
